requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
-e file:sodeep#egg=sodeep
-e file:.#egg=tme-i10n
absl-py==0.9.0            # via tensorboard, tensorflow
aiogram==2.6.1
aiohttp==3.6.2            # via aiogram
astor==0.8.1              # via tensorflow
async-timeout==3.0.1      # via aiohttp
attrs==19.3.0             # via aiohttp
babel==2.8.0              # via aiogram
beautifulsoup4==4.8.2
cachetools==4.0.0         # via google-auth
certifi==2019.9.11        # via aiogram, requests
chardet==3.0.4            # via aiohttp, requests
fire==0.2.1
gast==0.2.2               # via tensorflow
google-auth-oauthlib==0.4.1  # via tensorboard
google-auth==1.11.2       # via google-auth-oauthlib, tensorboard
google-pasta==0.1.8       # via tensorflow
grpcio==1.27.2            # via tensorboard, tensorflow
h5py==2.10.0              # via keras-applications
idna==2.8                 # via requests, yarl
joblib==0.14.1            # via scikit-learn
keras-applications==1.0.8  # via tensorflow
keras-preprocessing==1.1.0  # via tensorflow
markdown==3.2.1           # via tensorboard
multidict==4.7.5          # via aiohttp, yarl
numpy==1.18.1             # via h5py, keras-applications, keras-preprocessing, opt-einsum, scikit-learn, scipy, tensorboard, tensorflow
oauthlib==3.1.0           # via requests-oauthlib
opt-einsum==3.2.0         # via tensorflow
protobuf==3.11.3          # via tensorboard, tensorflow
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pytz==2019.3              # via babel
regex==2020.2.20
requests-oauthlib==1.3.0  # via google-auth-oauthlib
requests==2.23.0          # via requests-oauthlib, tensorboard
returns==0.13.0
rsa==4.0                  # via google-auth
scikit-learn==0.22.2.post1  # via sklearn
scipy==1.4.1              # via scikit-learn, tensorflow
six==1.12.0               # via absl-py, fire, google-auth, google-pasta, grpcio, h5py, keras-preprocessing, protobuf, tensorboard, tensorflow
sklearn==0.0
soupsieve==2.0            # via beautifulsoup4
tensorboard==2.1.1        # via tensorflow
tensorflow-estimator==2.1.0  # via tensorflow
tensorflow==2.1.0
termcolor==1.1.0          # via fire, tensorflow
toolz==0.10.0
typing-extensions==3.7.4.1  # via returns
urllib3==1.25.8           # via requests
werkzeug==1.0.0           # via tensorboard
wheel==0.34.2             # via tensorboard, tensorflow
wrapt==1.12.1             # via tensorflow
yarl==1.4.2               # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools==46.0.0        # via google-auth, markdown, protobuf, tensorboard
